extruder-failure-prediction/requirements.txt
pandas>=2.0.0
numpy>=1.21.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
